fix bug in sys_sysconf

+++ sysdeps/obos/generic/sysdeps.cpp
@@ -529,7 +529,7 @@ int sys_vm_unmap(void *pointer, size_t size)
     return sys_anon_free(pointer, size);
 }
 
-int sys_sysconf(int num, int* ret)
+int sys_sysconf(int num, long* ret)
 {
     return syscall2(Sys_SysConf, num, ret) == OBOS_STATUS_SUCCESS ? 0 : ENOSYS;
 }
